feat: 레이스 페이지 공동 순위 구현
