+++ app/src/main/java/com/rocket/cosmic_detox/data/datasource/season/SeasonDataSourceImpl.kt
@@ -3,6 +3,7 @@ package com.rocket.cosmic_detox.data.datasource.season
 import android.util.Log
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.Query
+import com.rocket.cosmic_detox.data.model.RankingInfo
 import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
 
@@ -15,8 +16,9 @@ class SeasonDataSourceImpl @Inject constructor(
             val seasonRef = firestore.collection("season").document(seasonId).collection("ranking")
             val querySnapshot = seasonRef.orderBy("totalTime", Query.Direction.DESCENDING).get().await()
 
-            val documents = querySnapshot.documents
-            val rank = documents.indexOfFirst { it.id == uid } + 1
+            val rankingList = querySnapshot.toObjects(RankingInfo::class.java).sortedWith(compareByDescending<RankingInfo> { it.totalTime }.thenBy { it.name })
+
+            val rank = rankingList.indexOfFirst { it.uid == uid } + 1
 
             if (rank > 0) {
                 rank
