+++ app/src/main/java/com/rocket/cosmic_detox/data/model/Ranking.kt
@@ -1,28 +1,9 @@
 package com.rocket.cosmic_detox.data.model
 
-
-private const val RANKING_TOP = "top"
-private const val RANKING_BOTTOM = "bottom"
-
-sealed class Ranking {
-    abstract val id: String
-}
-
-data class RankingTop(
-    val topItems: List<RankingInfo>
-) : Ranking() {
-    override val id: String = RANKING_TOP
-}
-
-data class RankingBottom(
-    val bottomItems: List<RankingInfo>
-) : Ranking() {
-    override val id: String = RANKING_BOTTOM
-}
-
 data class RankingInfo(
     val name: String = "",
     val point: Int = 0,
     val totalTime: Int = 0,
-    val uid: String = ""
+    val uid: String = "",
+    val rank: Int = 0
 )
