+++ app/src/main/java/com/rocket/cosmic_detox/presentation/view/fragment/race/adapter/RaceAdapter.kt
@@ -8,13 +8,13 @@ import androidx.recyclerview.widget.ListAdapter
 import com.rocket.cosmic_detox.R
 import com.rocket.cosmic_detox.databinding.ItemRankingBottomListBinding
 import com.rocket.cosmic_detox.databinding.ItemRankingTopListBinding
-import com.rocket.cosmic_detox.data.model.Ranking
-import com.rocket.cosmic_detox.data.model.RankingBottom
 import com.rocket.cosmic_detox.data.model.RankingInfo
-import com.rocket.cosmic_detox.data.model.RankingTop
 import com.rocket.cosmic_detox.presentation.view.fragment.race.RankingDividerItemDecoration
 import com.rocket.cosmic_detox.presentation.view.fragment.race.RankingItemClickListener
 import com.rocket.cosmic_detox.presentation.common.ViewHolder
+import com.rocket.cosmic_detox.presentation.model.Ranking
+import com.rocket.cosmic_detox.presentation.model.RankingBottom
+import com.rocket.cosmic_detox.presentation.model.RankingTop
 
 enum class RankingType(val type: Int) {
     RANKING_TOP(0),
